Index / Core file structure organization

+++ src/Opdex.Core.Infrastructure.Abstractions/Data/Models/PairEntity.cs
@@ -5,7 +5,7 @@ namespace Opdex.Core.Infrastructure.Abstractions.Data.Models
         public long Id { get; set; }
         public string Address { get; set; }
         public long TokenId { get; set; }
-        public decimal ReserveCrs { get; set; }
+        public ulong ReserveCrs { get; set; }
         public string ReserveSrc { get; set; }
     }
 }
